Move enet to libs
